Update m1.medium RAM.
m1.medium was created with 10240 MB of RAM.  Change this to 4096
to be more dev env friendly.

AWS uses a flavor with 8 GB by default (m4.large), but the libvirt
provider is using 2 and 3 GB VMs, so 4 GB should be enough.

We may just need to make flavor selection easier.