Make it easy to run other installer commands.
Add an easy way to generate the install config for customizing the
deployment and document this in the README.  I need to do this every
time to limit the master and worker nodes to fit my dev env, but it
may be useful for other things.

The change also makes it easy to run other commands from the
installer, but doesn't document those as examples yet.

In passing, also reorder the sourcing of files in this script, so that
it's easy to override values from ocp_install_env in your user config
file as needed.