Move image variables into common.sh
The recent version bump only updated get_rhcos_image.sh
so we'll calculate the image hash in 03_configure_undercloud.sh
with the old filename.

Probably best to move the variables to the common.sh, and I also
removed the lolcat/figlet usage as I need to run get_rhcos_image.sh
on a different box without all the dependencies installed, so lets
keep the requirements as minimal as possible.