Sync 06_run_ocp.sh with upstream CI
The openshift/installer removed support for ENV vars, which makes
terraform prompt for the missing information. In upstream CI, where
answering prompts is not an option, they heredoc the install-config.yml
from the env vars.

This commit updates 06_run_ocp.sh to do the same. It also names the vars
the same as in the upstream CI repo.