Run DNS server in container at BM server
For the DNS short term solution, the external DNS should be
configured with the relevant SRV and CNAME records.

In OCP-DOIT environment we'll run a  DNS Server at BM that will
play the external DNS role.
The Neutron DNS will point to this server as the upstream server,
and the BM server will point to the LAB  external DNS server.