WIP: Add scripts for setting up CI resources
We need a DNS server in our CI tenant that can have entries dynamically
added when a new job starts.

This commit adds an ignition config to create such a server using the
CoreOS image with coredns. This sets up coredns to watch for changes to
the local filesystem, as well as a simple python HTTP server that can
update the entries dynamically.

With this server running and some minor hacks on the CI job[1], I am
able to get to a fully automated e2e deploy on vexxhost.

[1] https://github.com/trown/installer-e2e/commit/c2a8b66122366c5bcb5b77d226b4a2999a57a457