Fix recreation of install-config.yaml
Currently 06_run_ocp.sh creates install-config.yaml only if
ocp folder is missing. But when the installer is started, the
file is removed from the system. It leads to the fact that on
the second attempt of deploying OCP install-config.yaml is not
generated and the deployment fails immediately.

This commit adds a check that the file exists, and generates it
again in case of its absense.