Add `get-image-commit` script
This script will create a link to the commit a given image was built
from.

It takes two arguments: the "release image" which contains links to all
the images in that release as well as the "cluster image" which is the
actual image we want to know about.

Example:

    $ ./get-image-commit registry.svc.ci.openshift.org/openshift/origin-release:v4.0 machine-api-operator

Will return:

    $ https://github.com/openshift/machine-api-operator/commit/273292272d2d11bd37ea818ed8791f93eb93c3cf