Add `oc-image-commit`
It's got the same functionality as `get-image-commit` but using `oc`
instead of `podman`.

This is faster because `oc` doesn't download the images -- it can read
the metadata directly. On the other hand, it does need the `oc` binary
and possibly being logged in so I'm keeping the old `get-image-commit`
too.